ALFREDOPS-120 Confirmation dialog on the button that destroys your Alfresco + cleanup

+++ src/main/ts/eu/xenit/care4alf/web/js/amps.ts
@@ -1,7 +1,7 @@
 /// <reference path="care4alf.ts" />
 
-care4alf.controller('amps', function($scope, $http) {
-    var load = function(){
+care4alf.controller('amps', function($scope, $http, $window: Window) {
+    var load = function() {
        $http.get("amps/").success(function(modules) {
            $scope.modules = modules;
        });
@@ -13,7 +13,10 @@ care4alf.controller('amps', function($scope, $http) {
         $http.post("amps/", module).then(load,load);
     };
 
-    $scope.clear = function(){
-        $http.delete("amps/").then(load);
+    $scope.clear = function() {
+        if ($window.confirm("Are you sure you want to delete all the amps in this Alfresco? " +
+                "This will cause any features implemented in these amps to stop working.")) {
+            $http.delete("amps/").then(load);
+        }
     }
 });
